+++ src/clientity/exc/__init__.py
@@ -1,0 +1,4 @@
+# ~/clientity/src/clientity/exc/__init__.py
+from .base import (
+    Error, ClientityError
+)